Moving to `nvim-snippy` for snippets
I don't have a lot of time to create intricate snippets, so I won't use
most of the features `LuaSnips` has. `nvim-snippy` is a much smaller
plugin which should (hopefully) create less maintenance. However, it
doesn't support the VSCode JSON snippet format, so I had to switch to
`vim-snippets` instead for pre-defined snippets.

Also, for some reason `LuaSnips` and `nvim-cmp` didn't show snippets
that didn't start with alphanumeric characters. Switching to this new
engine fixed the problem (somehow, not sure what the issue was in the
first place).

I made a comment on an issue in `cmp-luasnips` on a similar problem
explaining all this. I might move back if I can get some help to find
the root cause. For now, `nvim-snippy` seems good enough for my needs.