A few changes to plugins and some telescope filtering:
Things like co-pilot or nvim-cmp seem to run in a Telescope prompt. This
isn't necessary here, so these changes filter it out. Also, adding back
in a statusline plugin along with gitsigns. Still trying to go for a
minimal setup, but these things help make the editor feel more complete.